Moves the currency to the org settings, rather than setting it at the metric level based on comments from Jeremy.
