+++ packages/back-end/src/controllers/metrics.ts
@@ -304,7 +304,6 @@ export async function postMetrics(
     userIdColumn,
     userIdTypes,
     anonymousIdColumn,
-    currency,
   } = req.body;
 
   req.checkPermissions("createMetrics", projects?.length ? projects : "");
@@ -357,7 +356,6 @@ export async function postMetrics(
     regressionAdjustmentOverride,
     regressionAdjustmentEnabled,
     regressionAdjustmentDays,
-    currency,
   });
 
   res.status(200).json({
@@ -428,7 +426,6 @@ export async function putMetric(
     "userIdColumns",
     "userIdTypes",
     "timestampColumn",
-    "currency",
   ];
   fields.forEach((k) => {
     if (k in req.body) {
